Changed Alignment panel from Border layout to Box layout.
The Box layout provides a more compact arrangement of rows than the
Border layout. It also provides an unlimited number of subpanels in a
vertical stack. The Border layout is limited to a top, middle, and
bottom. Since future changes are likely to add additional rows (bias
settings, for example), the Box layout now makes more sense.