Closer to applying biases, but uses a more difficult approach
This version was planned to apply the biases inside run_swift.java.
So this version added another overloaded of align_files_by_name to
accept the bias_x and bias_y.

However, in retrospect, it might be easier to do this in the special
pairwise part of swift_gui.java where the propagated affine transforms
are currently applied. So the next commit might reverse much of the
code from this commit by applying the biases there.

Note that this somewhat violates the goal of wrapping all the image
processing in "run_swift" rather than in "swift_gui". But given the
tight time constraints, this is the quickest way to implement biases.