Made the biases effectively global ... when "enter" is pressed.
The bias values were designed to be set per image. However, it is also
useful to specify them as global values.

This commit uses the "actionPerformed" function to copy the value of any
bias (x or y) into all other "sections" when the "enter" key is pressed.
This approach preserves the local versions internally (for possible
future use) while making it convenient to change globally.