Finished the "remove_sci" function to temporarily patch the swim bug.
It turns out that swim cannot handle any scientific notation (5.3e-6).
This commit includes a function named "remove_sci" which takes a string
and parses it for anything that can be interpreted as a Java Double.
When found, those strings are converted to %.9f to be compatible with
swim's limitation.