Converted the second window from a button to a zoom/pan area.
Note that there's a problem because there's no way to pass the zoom/pan
area to the mouse scroll callback. The single window version got away
with making that a global variable inside app_window.py. However, this
creates a conflict when two zoom/pan areas are created. In this case,
the global "zpa" is the last one that had been created.