Created separate detection and handling of left/right zoom/pan windows
As mentioned in the previous commit, the Python version of zoom/pan
window used a global variable to store the zpa. This commit uses 2
global zoom/pan areas (zpaL and zpaR) which are assumed to be created in
that order (left first, then right). These are then stored along with
the drawing area itself. Then when the drawing area is passed to the
callback, it can be compared to these previously stored values to see if
the callback was originated from the left or right zoom/pan window.

This approach seems to work, but it would need to be manually extended
for each additional window. It might be possible to store the values in
a list so that any number of windows could be created. However, for now
this will be sufficient to get pyswift_gui up and running.