Created a zoom_window class as subclass of app_window.zoom_pan_area
The proper use of the app_window.zoom_pan_area should be to subclass
with additional functionality.

This code was originally written as non-object-oriented, and is slowly
being migrated to an object-oriented framework. For example, the
mouse_scroll_callback should have been a member function rather than a
module function. These changes can be made as needed.