Added minimal changes to run in Python 2.7 and support integration
The current swift_gui code was written for Python 2. Unfortunately, the
current pyswim code was written for Python 3. The only change needed
was the removal of the unicode characters (in the comments). They were
originally included as an inline string, and Python 2 objected. They
were then changed to be in comments (#) and Python 2 still objected.
So they were removed completely.

Also, the traditional "if __name__ == "__main__" code was added so that
it could run stand-alone or be included without running the test cases.