Added support for printing image coordinates with shift-click.
This change involved moving more of the app_window.py functions into the
zoom_pan_area class. Those functions had been global within the
app_window.py file, but make more sense within the class.

There is still a problem with dragging that is not being ignored with
the shift key, but careful holding of the mouse can avoid this problem.

This brings up the issue of possibly needing to "disconnect" signals in
GTK as well as connecting them. This hasn't been explored yet.