Added a run_alignment_callback as a stub demonstrating the data fields
The run_alignment_callback function currently prints all of the
alignment values stored in the project. This is intended to be a guide
for anyone wanting to run an actual alignment.

Note that the code already generates an "index_list" containing the
index values of those sections to be aligned. The base image name is the
one referenced by that index, and the adjust image name is the next one
in the list. All the other parameters are indexed with the base index.

The following code prints these values for the whole alignment list:

  print ( "Aligning sections " + str(index_list) )
  print ( "Destination Path = " + destination_path )

  for i in index_list[0:-1]:
    print ( "===============================================================================" )
    print ( "Aligning " + str(i) + " to " + str(i+1) + " with:" )
    print ( "" )
    print ( "  base                     = " + str(alignment_list[i].base_image_name) )
    print ( "  adjust                   = " + str(alignment_list[i+1].base_image_name) )
    print ( "  skip                     = " + str(alignment_list[i].skip) )
    print ( "" )
    print ( "  translation window width = " + str(alignment_list[i].trans_ww) )
    print ( "  translation addx         = " + str(alignment_list[i].trans_addx) )
    print ( "  translation addy         = " + str(alignment_list[i].trans_addy) )
    print ( "" )
    print ( "  affine enabled           = " + str(alignment_list[i].affine_enabled) )
    print ( "  affine window width      = " + str(alignment_list[i].affine_ww) )
    print ( "  affine addx              = " + str(alignment_list[i].affine_addx) )
    print ( "  affine addy              = " + str(alignment_list[i].affine_addy) )
    print ( "" )
    print ( "  bias enabled             = " + str(alignment_list[i].bias_enabled) )
    print ( "  bias dx                  = " + str(alignment_list[i].bias_dx) )
    print ( "  bias dy                  = " + str(alignment_list[i].bias_dy) )

Note that this version still shows both images in both windows.
Future versions will show the destination images as in the Java version.