Began the process of generating aligned images from originals.
The current plan is to have the alignment code generate the new aligned
images and add whatever image names it wants to be displayed along with
the base image. This gives the alignment code the ability to choose what
it wants to be displayed. It may display the current image along with
any number of aligned or produced images. It may even display
neighboring images. As long as they're added to the list of images to be
displayed, then the code should try to display them.

The same plan should work for any custom drawing. There will be an empty
list of "things" to draw for each original image. The alignment code can
add anything that needs to be drawn to that list. It is expected that
these will be relatively simple primatives (lines, ellipses, etc) that
can be easily rendered by any drawing platform. The coordinate system
will be in the same "pixel" space as the image itself.

Much of this hasn't been implemented yet, but it should be easy enough.