Began sketching out some classes to hold annotated images.
An annotated image might have the image file name and a list of graphics
items to be drawn on that image (in the pixel coordinates of the image).

An alignment might have certain expected images (such as base and adjust),
but should also have a list of other arbitrary images to be displayed as
generated by the alignment algorithm. These might also include flags to
control the display of various parts of the annotated image (raw pixels,
border, coordinates, graphics items, named items, etc).