Added an image_list to the alignment class with "red_" prefix to test
An "alignment" contains the base image, parameters for the alignment,
and possibly some number of other images to be displayed.

These "other" images are contained in the "image_list" within the
alignment object. Each such image is actually an "annotated_image" which
means it can also contain annotations (points, rectangles, text...).

The matching of "other" images to viewing windows is done via an index
number in the zoom_window itself. For the current implementation, a
value of -1 means the base image. Any other number will be assumed to be
an index into the image_list.

For this prototype version, the image list is populated using the base
image names prefixed with "red_" to match some red images that had been
manually produced. Ultimately, it will be the alignment code that will
generate the new images and add them to the list.