Added "fake" alignment which just generates lighter and darker images
This version uses the GTK "saturate_and_pixelate" function to make some
visible change to the original images and save them in the destination
directory. This version creates a "light_..." and "dark_..." version of
each image. These are added to each alignment and can be displayed by
adding additional windows.

Note that a future enhancement might provide a way to show all images
that have resulted from the processing by determining the longest image
list among the various sections.