Removed the "align" image window which is replaced by extra windows
The previous version of the code had an original image, an "aligned"
image, and a list of additional images that had been produced by the
alignment code itself. However, the hard-coding of the "aligned" image
isn't needed since the alignment code can produce that image (and its
associated window) as needed. Removing that hard-coded image (and its
window) simplifies the code. In fact, it would be possible to include
even the original image in that list. This wasn't done at this time
because the original image does have a somewhat separate nature.