Added code to generate windows as needed for "Refresh" menu option.
When an alignment is run, each alignment may contain any number of
additional images in its own image_list. The "Refresh" option now
examines each alignment to see how many images are in its list and
allocates the proper number of windows to display the maximum in any
of the alignments. Images are stored and displayed by their location
in the list.