PySwift: Began implementing coordinate system concepts
There are 3 somewhat natural coordinate systems that might be useful:

  1) Pixel Coordinates: These are the raw screen coordinates
  2) Image Coordinates: These are relative to the image displayed
  3) Scaled Coordinates: These are a fraction of the screen size

These are flagged by the single characters 'p', 'i', and 's'.

Pixel coordinates are useful for direct drawing where the code may
draw in screen coordinates or manage transformations directly. The code
may or may not zoom the drawing with the window as needed.

Image coordinates will be most useful for annotating images. Any
graphics drawn in this coordinate system will scale and shift with the
images.

Scaled Coordinates are useful for drawing things that should scale with
the window size itself. These are typically status items like SNR.

All graphics primitives will accept a coordinate system in their
constructors.