PySwift: Added support for most planned kinds of annotations.
This version supports points, lines, rectangles, and text. They can be
added to each image as it is being aligned.

This version supports both pixel ("p") and image ("i") coordinate modes
as described in the earlier commit message.

These annotations are intended to make it easy for recipe developers to
add annotations to the images being produced.