Fixed fields to data bug when not scrolling.
The code copies the fields into the data and then the data into the
fields when the scroll wheel changes layers. However, any changes to
fields that are not followed by a scroll will not be reflected in the
data. This commit copies the fields into the data whenever an alignment
is started to be sure that it's using the latest field data even if
there has been no scrolling.