Began grafting on the point picking functionality (not done yet).
The ability to pick points was not anticipated in the original design.
However, there are some features of the design that can lend itself to
the task. This commit adds a mew "marker" type of graphic that can be
added to images. It's intended to allow marking of corresponding points
between the "sta" and "mov" images, so it is restricted to just those
two panels.

The point picking mode is enabled through the menu, and it will turn the
cursor into a "hand" when in that mode. While both the "sta" and "mov"
windows will respond to point picking, only the "mov" window has the
graphic list needed to store and display the points because it's an
"annotated image". The base image was created before the annotation
idea had been invented, so it will need to be upgraded before point
picking will work.

There are likely to be a number of other problems, and this code is
getting messier and messier with each "quick prototype".