Display the base image as part of the image list for each layer.
The previous versions had special code to handle the base image and draw
it in its own window.

However, this approach always ended up putting the base image in the
first position.

The solution is to require that the base image be added to the image
list in order to be displayed. It can be added in any position (or no
position at all) as desired by the alignment code. The default (when
loading images) will be to place the newly loaded image in the first
slot (with all other slots empty). The alignment process is then free
to remove it and add it as needed along with other images.