Implemented the new alignment layer system.
This version uses the new alignment layer system where the original
(base) images form the viewable stack. All images are added to each
visible layer by the alignment process (engine?). This includes the
original image so that it may be placed in any order as desired by
the alignment engine.

In this current implementation, the alignment engine places the
original image in the center panel. The next lower image in the stack
is placed in the left panel (when it exists), and the corresponding
aligned image is placed in the right panel when it exists.

Note that the manual point selection in this version is broken
because it was based on the previous windowing arrangement.