Tom's updates to align_swiftir.py and partial work on point picking.
The change to align_swiftir.py includes an alignment_process class to
hold all of the data for a SWiFT-IR alignment.

The current changes to pyswift_gui.py have been the minimal to use the
new changes. The current pyswift_gui does not yet take advantage of the
new features.