Added a dummy empty image to skipped layers to align originals in col 2.
Importing images creates the first column of images in a project. All
other images are added as a result of processing. These additional
images were intended to be added to the right of the original image to
reflect their sequence of generation. However, the current alignment
"engine" clears the list and adds the previous layer's image before the
original image for the current stack. But that only happens when the
alignment engine operates on that layer. Skipped layers are currently
not operated on, and so the images in that layer are not reordered which
leaves the original image in column 1 rather than column 2. This quick
fix simply inserts a dummy empty image into column 1 so that the
original image will appear in column 2.

Note that using named columns could solve this problem in the future.