Fixed point picking that was broken by earlier redesign of layers
Commit b3bb42805dbb750c0ae63484a26cdd653b848bdb reorganized the layering
system and broke the point picking as expected.

The point picking depends on certain windows in certain positions. These
were changed during the earlier redesign, but the point picking system
hadn't been updated so other issues could be addressed. This was
mentioned in commit b3bb428.

Note that this version relies on the images being in the first two
window positions. This is a very fragile design and should be redone
after the "proof of concept" version is complete.