Converted some of the menu items to checkbox menu items.
Some of the menu items are modes that can be toggled. The earlier
version just used regular menu items to toggle them without showing
the current state. Some of these will now show state. Note that these
were all implemented assuming the default to be off and not checking
the actual state. That's a shortcut for now.