Began working to pull user-selected points out of image lists.
The points are in image coordinates (origin at upper left) and are
stored as "graphic_marker" annotations in the "graphics_items" list for
each image. The "graphic_marker" class has an "index" item which will be
set to 0 when matched with the next image and 1 when matched with the
previous image (it might be better to use -1 for prev, and 1 for next).

These should be sufficient to determine the alignment, but it's
complicated by the fact that images are only known by position in the
list and don't have any designated roles yet. Adding roles would help.