Changed zoom_panel's "name" field to be a "role" field.
The plan is for each panel to have a "role". Each "layer" in the stack
will contain a dictionary of annotated images keyed by their "role". The
framework will display images in panels with matching roles. Either the
images or the panels will also have flags such as "pick points" to
signal additional capabilities.