More work on shadowing the image_list with the image_dict (not done yet)
The use of the image_dict will allow pairing of window roles with image
roles (keys to the images in a layer). This will eliminate the confusing
use of special indexes which imply image roles. This is only partially
done at this commit.

Eventually (after the shadowing is complete), the image_list should be
completely removed.