Automatically center images on open and align.
This commit includes a lot of "overkill" in the "center_all_images"
function. It turned out that the problem wasn't in "center_all_images",
but in the way that GTK creates its panel structure. Many of the panel
features simply aren't ready to be drawn until the first "expose_event"
is sent. So all attempts to get the window size were returning 1x1 until
the first "expose_event" was received. So much of the centering code was
already working as additional efforts were made. What ended up actually
solving the problem was to call the "center_all_images" function from
inside the expose_event callback, and to trigger it with a flag that
gets set externally when new images are loaded.

If this were more than a prototype, the "center_all_images" function
(and related code) would deserve a careful re-write. That can wait.