Adjusted the text positions for lab computers and center on load.
The text positioning was done by trial and error to get something that
fit well. However, different computers appear to have different default
font widths. Proper text positioning would involve sizing of the string
pixels with the font, but hard-coded values are fine for this prototype.

Also, the centering from the previous commit was added when loading new
images into the project.