Added code to support aligning by Swim Window or Match Points.
This commit adds a new gtk.ComboBox to each alignment layer. The new
field can be in either of these states:

  0 = Swim Window
  1 = Match Point

The value of the field can be detected in the aligment loop for each
layer to direct the processing.

Note that this field is not being stored in the JSON project file and
neither are the alignment points which can be selected. All of these new
values are not being used by any alignment processes yet (although they
are available withing the run_alignment_callback function).