Change "Align Forward" to include current layer in count.
This change causes "Align Forward 1" to align the currently displayed
layer and not the subsequent layer.

This is somewhat inconsistent with the use of "Set Forward 1" which
will move the current parameters into the subsequent layer. However,
it was felt that the ability to align (and re-align) the layer in view
was an important use case. At some point it may be worthwhile to rename
these options so that they make sense.