Fixed the need to scroll to the end before aligning.
The previous version needed to be scrolled through all of the images
before performing an alignment. This was true because the fields:

  self.align_method
  self.align_method_text

were not defined in the constructor, but they were defined by being
swapped in and out of the GTK widgets that displayed them. Scrolling
through the stack one time performed this initialization.

This fix defines them in the "alignment_layer" constructor.