SWiFT-Specific: Automatically load reference images into "ref".
When importing images, they pretty much belong only in the "base" slot.
However, this means that the alignment code needs to run in order to
load images into the "ref" slots (because even having a "ref" slot is
specific to the alignment method being used). This means that images
cannot be annotated (point picking, for example) until after the
alignment code has been run to provide the "ref" image. So this commit
adds special code to clone the "base" images into the "ref" slots for
the appropriate layer. So point picking can be done without running
an actual alignment.