Since the base and ref are now created on load, do not recreate in align
The earlier concept was that the framework would only load base images.
Any additional images loaded would be the responsibility of the alignment
method. However, to make it simpler, the code now automatically loads both
the "base" images and the "ref" images. Since they are already loaded, the
alignment code doesn't need to create them. This has the added advantage
that any alignment points (which are only defined in the "base" and "ref"
images) will remain untouched by an alignment. So the alignment process
doesn't have to transfer them when recreating those images.