Checked for cases where the image was empty (had shown parent dir).
When an image file name was an empty string, the code had converted it
to an absolute path and then displayed the last section of it (which
turned out to be the parent directory) as the file name.

This commit checks for an empty file name and then generates a new
annotated image from "None" when detected.