This version reads the cumulative_afm from the JSON file and uses it
The "cumulative_afm" has been stored in the "method_results" key of the
"align_to_ref_method" dictionary in the JSON file. This commit adds the
code needed to read it back when loading a JSON project file. Note that
this code also has a commented section of code that was used to verify
that the "prev_afm" that had been obtained from align_proc matched the
"cumulative_afm" stored in the "results_dict" to be sure there were no
rounding effects due to possibly different representations. The "exit"
was never triggered and is now commented out (soon to be removed).