Added support for deleting alignment points.
This version has a menu option to delete points when in point mode.

Because the points are paired with each other from the front of each
list onward, and because new points are always added to the end of the
list, it's usually best to delete points in matching pairs and replace
them in matching pairs.

Also note that points are color-coded from the beginning of the list for
each image. So deleting points that are not at the end of the list will
cause other points to change colors.