Updated the test images to show the image number in the image itself.
These images contain pixel drawings of the numbers 1-7 in the center of
the image. The pixel drawings are fuzzy and light to minimize any impact
on the alignment itself. These are handy for actually seeing which image
is being shown without any ambiguity. They are especially handy for test
images.