Show proper ref image and support match points across skip.
Note that this version makes no effort to preserve match points across a
skip when a skip is changed. They are "officially" in an undefined state
and should be cleared and added from scratch. They may be found to be
correct, and may be used in that case. But it hasn't been designed into
the code or tested.