Added logic to automatically jump to unskipped images on menu change.
The previous version would continue to show a skipped image even when
the "Show Skipped" was turned off ... until the scroll wheel was moved.
This happened because the filtering logic was only triggered via the
mouse scroll wheel and the "move_through_stack" function. This version
also triggers the filtering when the menu option itself is changed.

This was accomplished by adding another case to the "move_through_stack"
function. The previous version could only move up or down. This version
can take a movement parameter of 0 which is no movement. So when the
"show_skipped_layers" flag is false, the "move_through_stack" function
will now find the next closest unskipped image in either the forward
direction (movement>0), backward direction (movement<0) or the closest
unskipped image in either direction (movement==0).