Added an "SNR Skip" mechanism to flag layers that fall below threshold.
The current "run_alignment_callback" isn't designed to dynamically skip
alignment jobs based on SNR values. However, it is easy to use the SNR
values to set a flag on each layer falling below a threshold. Then those
flags can be easily copied into the actual "skip" field for the next run.

This version implements an "snr_skip" flag that's set as each alignment
is processed. The SNR skip flag is shown as a yellow "X" (rather than a
red "X" for actual skips). There is a button that can copy all of the
"SNR skip" flags into the normal "skip" flags for a second pass to do
the actual skipping.