Quick fix for Windows path separators and smaller default window size.
Windows paths typically contain "backslash" separators which are seen as
"escape" sequences to Python's JSON parser. But Windows also seems to
accept "normal" forward slashes as path separators as well. This quick
fix simply changes the back slashes to forward slashes when writing
paths. Note that this also needed to be done in the "annotations" since
any "backslashes" (whether used as paths or not) will cause the JSON
parser to fail.

This version also changed the global window size defaults to be 600
rather than the previous 800. Unfortunately, the default also becomes
the smallest that the window can be resized, so this smaller size can
allow the program to more easily used on smaller screens while adding
only a minor inconvenience (resizing) to those with larger screens. At
some point it would be good to try to allow the windows to be scaled
down as well as up.