Added the infrastructure to write a project file and call (exec) runner
The code to write a JSON project was moved from the menu code into its
own function to be called by both the menu code (on demand) and the run
code (to export).

A simple "run_json_project.py" program was added to eventually run the
JSON project exported by pyswift_gui.py. Eventually this might be a user
selectable file name to allow running all kinds of variants, but for now
the name "run_json_project.py" is hard-coded.

This version of "run_json_project.py" doesn't do anything other than
announce that it has started and exited.