Modified run_jason_project.py to print most of the JSON fields.
The "run_json_project.py" program was adapted from the code that reads a
JSON project into the internal structures inside "pyswift_gui.py". The
code that runs a JSON project need not (and maybe should not) use the
same exact structures that are used by the GUI. So this version was
created to be completely independent of the code in pyswift_gui.
Instead, it uses Python's internal JSON library to convert the JSON into
the familiar dictionary and list structures native to Python. If the
implementation of a particular runner needs different data structures,
then that should be part of that runner's internal code.

Note also that right now there is only one "external" runner, and its
name is hard-coded to "run_json_project.py". That should be enough to
get us started writing code for the various job management systems that
we want to target. It shouldn't be too much trouble to make the choice
of a "runner" easily selectable in the future.