Changed to use "__import__" linkage rather than "exec" of string.
The previous version just called "exec" to execute "run_json_project",
but there are many problems with that approach. This version uses the
"__import__" function followed by executing the "main" function inside
the imported module.

This version of "run_json_project.py" also imports from "pyswift_gui" to
obtain the "graphic_primitive" class. It's not so clear whether that's a
good idea or not. It's handy because uses code already available in the
pyswift_gui module, but it also creates a coupling that maybe shouldn't
be there if we want to use the JSON as the pure interface without any
other dependency.