Duplicated the actual running code from pyswift_gui into run_json_project
This version does isolate the running code. However, it imports all of
the important structures from pyswift_gui and accepts the data for those
structures as call parameters. While it still retains the "main"
function that reads from the JSON file, the "run_alignment" function
doesn't call that "main" function or use that data.

The difficulty is that the process of "running" the alignment has also
done the updating of the graphics for the images. It's not clear how
that data should come back from a more isolated "runner" in another
thread or process or machine.