Moved functionality from run_json_project.py to pyswift_run_external.py
The run_json_project was originally intended to run directly from JSON
input only. However, this would lose considerable functionality in being
able to access the internal data structures in pyswift_gui. So the
runner code from pyswift_gui was added to run_json_project during the
development phase. Now that it seems to work, that functionality has
been removed from run_json_project.py and added to the new file
pyswift_run_external.py.

The current plan for multiple "runners" is to prefix each runner's file
name with "pyswift_run_". These can then be found and dynamically loaded
at startup to make them available in pyswift_gui. This will help work
out the proper interfaces between the GUI and the runners.

It is anticipated that "run_json_project.py" will be renamed to follow
that convention, however, since it was also changed in this commit, it's
clearer to do the rename as a separate commit step.