Updated pyswift_gui.py to find modules starting with "pyswift_run_"
This commit searches the current directory (should search the __FILE__
directory) for any other files named "pyswift_run_*" and adds them to
the list of alignment runners. Note that in the CellBlender terminology,
these are really engine/runner pairs, but have been called "runners" for
short. A new "runner" named "pyswift_run_threads" has also been added as
a test. It's mostly a copy of pyswift_run_json.

This commit does not actually execute those new runners yet since the
standard interface has not yet been defined. But it does provide the
functionality to find new "runners" and add them to the list of choices
when the program starts.