Returned pyswift_run_json to using dictionary references
The "JSON" runner might be expected to be coded using simple dictionary
and list references similar to JSON itself. That was the original
design, but it was modified with set and get functions to look more
"pythonic". That more "pythonic" version is now preserved in the newly
added "pyswift_run_threads.py" file, so pyswift_run_json is being
returned to the dictionary-like reference system.