Modified the JSON and Threads runners to take "standard" parameters.
The "run_alignment" function is currently defined as:

  def run_alignment ( project_file_name,
                      align_all,
                      alignment_layer_index,
                      num_forward,
                      snr_skip,
                      snr_halt ):

This is now implemented in both the JSON and Threads runners. It is also
passed for all "generic" runners by pyswift_gui.

At some point (soon) it would be good to make these named parameters so
that not all runners will require identical signatures. The runner
modules might even support a "capabilities" functionality.