Commented out the blanking of any skipped reference images
When testing the JSON runner, the reference image for skipped images was
blank. This was a result of line 1218 creating a blank image for the
"ref" with the "annotated_image" constructor. It's not clear why this
had been done. That image might be recreated by the alignment code with
the internal runner (which assigns images to layers as they are
created).

This points out a problem yet to be solved. The runners create images on
disk, but they don't currently have a way to insert them into the stack
as the internal runners can do.