Added a "temp" flag to annotations that shouldn't be written to JSON
The "File Size" annotation is now added to files that exceed the
currently set file size limit. However, these annotations depend on the
max_image_file_size parameter that can be set before loading the JSON
file. So images that were "too big" when opened with one setting may no
longer be "too big" when opened with another setting.

There are a number of ways to have dealt with this, but this commit
defines a "temp" flag in all annotations. Any annotations generated with
this "temp" flag are simply not written to the JSON file. Currently, the
only annotation constructor that accepts this flag is "graphic_text" as
needed for the "File Size" annotation mentioned above. The "temp"
argument to the constructor is defaulted to False to be consistent with
all earlier behavior.