Changed max image file size limit from 16MB (testing) to 700MB
Our "large" images are currently around 600MB, so the testing limit of
16MB would not load any of these larger images. The new limit of 700MB
will allow these larger images unless reduced through the "Set" menu.

Different users on different machines may choose different settings, so
it's not clear what the best default should be. This larger default was
chosen so that anyone working with those larger images would not be
surprised by the new behavior. Eventually, the loading will be more
dynamic and this file size limit shouldn't be needed.