Added upgrade support for JSON version 0.2 (adding "scales")
The new JSON format (0.2) will add a "scales" dictionary to the "data".
The "alignment_stack" (previously directly in "data" dictionary) will
now be duplicated under each of the "scales" keys (1, 2, 5, 10, 32...).

This commit reads either format and upgrades as needed to produce the
current format internally. It also reads from the current format. But it
still uses the single alignment stack internally (does not yet support
multiple scales).

This upgrade approach is very similar to the data model upgrading used
in CellBlender for many years. Just as with CellBlender, the upgrading
is relatively transparent to the user. An upgrade warning is issued:

  Upgrading model from 0.1 to 0.2
  Saving will use the new format.

As the warning describes, saving will use the new format. So to retain
the current format, either save to a different file or don't save.