Began adding actual support for multiple scales.
This version can create multiple scales, and each scale is like a
separate project. Each can be aligned separately at this point. There
is no logical link between scales, and the images must currently be
imported separately (as if they were separate projects).

Note also that the multiple scales are not yet stored in the JSON or
read from the JSON. Only the currently selected scale is stored to
the JSON file when the project is saved.