Added 4k x 4k versions of the original vj_097 images for scaling testing
The current repository contains a set of small (807x780) images used for
testing. These images are too small for meaningful testing of several
scales, so this commit adds "upsized" versions (4238x4096) which can
then be "downsampled" to get a range of sizes.

These new files have the same name with "_4k4k" added, and they are in
the tests/vj_097_4k4k subdirectory in the project.