Added a separate "Scaling" menu item to hold scaling operations.
The scaling operations were previously under the Images menu, but that
was not an obvious location. So a new "Scaling" menu bar option was
added to support the scaling operations. Note that this is different
from the "Scale" menu bar option which is just selecting which scale to
view and work with. The "Scale" menu bar list is generated dynamically
based on the settings in the "Scaling" menu.