Changed to link scale 1 images rather than copy (posix only)
This commit detects the special case of scale 1 to either link or
directly copy images into the "scale_1" directories. This hasn't yet
been tested on Windows to ensure that the shutil.copyfile() works.

The previous version used align_swiftir.swiftir.scaleImage in all cases.
This was expensive for large images.

Note that it's recommended that all projects include '1' as one of the
scales. This may not be needed eventually, but is wise at this point.