Added code to support plotting functions.
This is a preliminary implementation using hard-coded plotting code to
show the x and y offsets across the stack.

The current version contains the ability to store custom code in the
JSON project file, but does not yet execute that code. The code is
stored as base64 encoded pickle of the actual code strings. This was
done to ensure that the resulting string would be compatible with JSON
encoding. The raw pickle may still contain both single and double quotes
which will cause problems with JSON. So the extra base64 encoding solved
that problem. The final string was split into a list of strings of 40
characters each. A final empty string ("") was added to avoid dealing
with the "trailing comma" issue.

This version shows any matplotlib windows in the current thread and as
a result it blocks until the matplotlib window is closed. It would be
better to run in either a separate thread or process. But this works for
now and creates a framework for custom plotting.