Added execution of dynamic plotting code via the PlotCode Dialog
This version includes a somewhat clunky support of dynamic plotting
execution. The "Show/Plot Code" option will bring up a dialog box
containing either the code found in the project or the default plotting
code. The "OK" button should be pressed just once after the first time
that dialog is opened. Afterward, changes can just be made to the text
and they will be executed with the next "Show/Plot" option. Even though
the PlotCode dialog contains a close button, it should NOT be closed or
it can't be opened again.

To avoid any problems (for now), follow this procedure:

   Load the project as normal
   Use the "Show/Plot" option to use the default plotting (or saved)

To modify the plotting, follow this procedure:

   Load the project as normal
   Use the "Show/Plot Code" option to bring up the dialog
   Click the "OK" button once
   Edit the code as desired
   Execute the plot with the "Show/Plot" option
   Continue to Edit or Execute as desired

Alignments may be done, and scales may be changed. The current scale
will be available to the code to plot code via the JSON model which
is available as the "d" variable to the plotting script. The default
script shows how to navigate the JSON model and find the current scale
to use for selecting what to plot.